Put shared work in XFiles (old XFilesDana module)
